[Gold V] Title: 옥상 정원 꾸미기, Time: 8 ms, Memory: 2416 KB -BaekjoonHub

+++ 백준/Gold/6198. 옥상 정원 꾸미기/옥상 정원 꾸미기.cc
@@ -18,7 +18,7 @@ int main() {
     cin >> n;
 
     for (int i = 1; i <= n; i++) {
-        long long height;
+        int height;
         cin >> height;
 
         while (!stk.empty()) {
